2839: advanced search form updates, add form to public search page, submit sequence for docket number search
