+++ web-client/src/presenter/actions/AdvancedSearch/validateCaseDocketNumberSearchAction.js
@@ -1,0 +1,33 @@
+import { isEmpty } from 'lodash';
+import { state } from 'cerebral';
+
+/**
+ * validate case advanced search form
+ *
+ * @param {object} providers the providers object
+ * @param {object} providers.applicationContext the application context
+ * @param {Function} providers.get the cerebral get function
+ * @returns {Promise} async action
+ */
+export const validateCaseDocketNumberSearchAction = async ({ get, path }) => {
+  const searchForm = get(state.advancedSearchForm);
+  const errors = {};
+
+  if (!searchForm.docketNumber) {
+    errors.docketNumber = 'Enter a docket number';
+  }
+
+  const isValid = isEmpty(errors);
+
+  if (isValid) {
+    return path.success();
+  } else {
+    return path.error({
+      alertError: {
+        messages: Object.values(errors),
+        title: 'Please correct the following errors:',
+      },
+      errors,
+    });
+  }
+};
