+++ web-client/src/presenter/sequences/submitCaseDocketNumberSearchSequence.js
@@ -1,0 +1,27 @@
+import { caseExistsAction } from '../actions/caseExistsAction';
+import { navigateToCaseDetailAction } from '../actions/navigateToCaseDetailAction';
+import { set, unset } from 'cerebral/factories';
+import { setAlertErrorAction } from '../actions/setAlertErrorAction';
+import { setDocketNumberFromAdvancedSearchAction } from '../actions/AdvancedSearch/setDocketNumberFromAdvancedSearchAction';
+import { setValidationErrorsAction } from '../actions/setValidationErrorsAction';
+import { state } from 'cerebral';
+import { validateCaseDocketNumberSearchAction } from '../actions/AdvancedSearch/validateCaseDocketNumberSearchAction';
+
+export const submitCaseDocketNumberSearchSequence = [
+  validateCaseDocketNumberSearchAction,
+  {
+    error: [
+      setAlertErrorAction,
+      setValidationErrorsAction,
+      unset(state.searchResults),
+    ],
+    success: [
+      setDocketNumberFromAdvancedSearchAction,
+      caseExistsAction,
+      {
+        error: [set(state.searchResults, [])],
+        success: [navigateToCaseDetailAction],
+      },
+    ],
+  },
+];
