+++ web-client/src/styles/custom.scss
@@ -1391,8 +1391,13 @@ button.change-scanner-button {
     }
   }
 
+  .advanced-search__form-container {
+    height: 326px;
+  }
+
   .advanced-search__button {
     margin-top: 39px;
+    margin-bottom: 20px;
   }
 }
