+++ web-client/src/views/PublicSearch.jsx
@@ -1,15 +1,7 @@
 import { BigHeader } from './BigHeader';
 import { Mobile, NonMobile } from '../ustc-ui/Responsive/Responsive';
+import { SearchForm } from './AdvancedSearch/SearchForm';
 import React from 'react';
-
-const SearchInformation = () => (
-  <p>
-    Anyone can search for a case in our system for cases filed{' '}
-    <span className="text-semibold">on or after May 1, 1986</span>. If you
-    aren’t affiliated with that case, you will only see limited information
-    about that case.
-  </p>
-);
 
 export const PublicSearch = () => {
   return (
@@ -18,7 +10,7 @@ export const PublicSearch = () => {
         <BigHeader text="Search for a Case" />
 
         <section className="usa-section grid-container advanced-search">
-          <SearchInformation />
+          <SearchForm />
         </section>
       </NonMobile>
       <Mobile>
@@ -26,7 +18,7 @@ export const PublicSearch = () => {
 
         <section className="usa-section grid-container advanced-search">
           <h2>Search for a Case</h2>
-          <SearchInformation />
+          <SearchForm />
         </section>
       </Mobile>
     </>
